sync

+++ tools/pecl-package.php
@@ -6,48 +6,52 @@
 # role="doc" => in $(pecl config-get doc_dir), which is /usr/share/doc/pecl/swoole on RPM distro (LICENSE being an exception, manually moved to /usr/share/licenses)
 # role="test" => in $(pecl config-get test_dir), which is /usr/share/tests/pecl/swoole on RPM distro
 
-define('SWOOLE_COLOR_RED', 1);
-define('SWOOLE_COLOR_GREEN', 2);
-define('SWOOLE_COLOR_YELLOW', 3);
-define('SWOOLE_COLOR_BLUE', 4);
-define('SWOOLE_COLOR_MAGENTA', 5);
-define('SWOOLE_COLOR_CYAN', 6);
-define('SWOOLE_COLOR_WHITE', 7);
-
-function swoole_log(string $content, int $color = 0)
-{
-    echo $color ? "\033[3{$color}m{$content}\033[0m" : $content;
-}
-
-function error_exit(string ...$args)
-{
-    foreach ($args as $arg) {
-        swoole_log('ERROR: ' . $arg, SWOOLE_COLOR_RED);
-    }
-    echo "\n";
-    exit(255);
-}
+require __DIR__ . '/functions.php';
 
 function check_source_ver(string $expect_ver, $source_file)
 {
-    static $source_ver_regex = '/(SWOOLE_VERSION +)("?)(?<ver>[\w-.]+)("?)/';
+    static $source_ver_regex = '/(SWOOLE_VERSION +)("?)(?<ver>[\w\-.]+)("?)/';
     $replaced = false;
     _check:
     $source_content = file_get_contents($source_file);
-    if (!@preg_match($source_ver_regex, $source_content, $matches)) {
+    if (!preg_match($source_ver_regex, $source_content, $matches)) {
         swoole_log(
             "Warning: Match SWOOLE_VERSION Failed, skip check!\n",
             SWOOLE_COLOR_MAGENTA
         );
         return;
     }
+
     $source_ver = $matches['ver'];
+
+    // auto fixed sub version values
+    if (strpos($source_content, 'SWOOLE_MAJOR_VERSION') !== false) {
+        $version_parts = array_values(array_filter(preg_split('/(?:\b)|(?:(?<=[0-9])(?=[a-zA-Z]))/',
+            $source_ver), function (string $char) {
+            return preg_match('/[0-9a-zA-Z]/', $char);
+        }));
+        list($major, $minor, $release, $extra) = $version_parts;
+        $source_content = preg_replace(
+            '/^(\#define[ ]+SWOOLE_VERSION_ID[ ]+)\d+$/m',
+            '${1}' . sprintf('%d%02d%02d', $major, $minor, $release),
+            $source_content
+        );
+        (function (&$source_content, $replacements) {
+            foreach ($replacements as $replacement) {
+                $regex = '/^(\#define[ ]+SWOOLE_' . $replacement[0] . '_VERSION[ ]+' . (is_numeric($replacement[1]) ? ')\d+()$' : '")[^"]*("$)') . '/m';
+                $source_content = preg_replace(
+                    $regex, '${1}' . $replacement[1] . '${2}',
+                    $source_content,
+                    1
+                );
+            }
+        })($source_content, [['MAJOR', $major], ['MINOR', $minor], ['RELEASE', $release], ['EXTRA', $extra]]);
+        file_put_contents($source_file, $source_content);
+    }
+
     if (!preg_match('/^\d+?\.\d+?\.\d+?$/', $source_ver)) {
         $is_release_ver = false;
-        swoole_log(
-            "Notice: SWOOLE_VERSION v{$source_ver} is not a release version number in {$source_file}\n",
-            SWOOLE_COLOR_YELLOW
-        );
+        swoole_warn("SWOOLE_VERSION v{$source_ver} is not a release version number in {$source_file}.");
     } else {
         $is_release_ver = true;
     }
@@ -56,13 +60,17 @@ function check_source_ver(string $expect_ver, $source_file)
         case -1: // <
             {
                 if ($replaced) {
-                    error_exit("Fix version number failed in {$source_file}\n");
+                    _replaced_error:
+                    swoole_error("Fix version number failed in {$source_file}");
                 }
-                swoole_log(
-                    "Notice: SWOOLE_VERSION v{$source_ver} will be replaced to v{$expect_ver} in {$source_file}\n",
-                    SWOOLE_COLOR_YELLOW
+                swoole_warn("SWOOLE_VERSION v{$source_ver} will be replaced to v{$expect_ver} in {$source_file}.");
+                $source_content = preg_replace(
+                    $source_ver_regex, '$1${2}' . $expect_ver . '$4',
+                    $source_content, 1, $replaced
                 );
-                $source_content = preg_replace($source_ver_regex, '$1${2}' . $expect_ver . '$4', $source_content, 1);
+                if (!$replaced) {
+                    goto _replaced_error;
+                }
                 file_put_contents($source_file, $source_content);
                 $replaced = true;
                 goto _check;
@@ -71,34 +79,41 @@ function check_source_ver(string $expect_ver, $source_file)
         case 1: // >
             {
                 if ($is_release_ver) {
-                    error_exit("Wrong SWOOLE_VERSION {$source_ver} in {$source_file}\n");
+                    swoole_error("Wrong SWOOLE_VERSION {$source_ver} in {$source_file}, please check your package.xml.");
                 }
             }
             break;
     }
 }
 
+// all check
+swoole_execute_and_check('php ' . __DIR__ . '/config-generator.php');
+swoole_execute_and_check('php ' . __DIR__ . '/arginfo-check.php');
+swoole_execute_and_check('php ' . __DIR__ . '/code-generator.php');
+swoole_execute_and_check('php ' . __DIR__ . '/constant-generator.php');
+swoole_execute_and_check('php ' . __DIR__ . '/build-library.php');
+swoole_execute_and_check('php ' . __DIR__ . '/phpt-fixer.php');
+
+// prepare
+swoole_ok('Start to package...');
 $this_dir = __DIR__;
 $tests_dir = __DIR__ . '/../tests/';
 `cd {$tests_dir} && ./clean && cd {$this_dir}`;
 
-$root_dir = __DIR__ . '/../';
+$root_dir = SWOOLE_SOURCE_ROOT;
 
 // check version definitions
-$package_ver_regex = '/<version>\s+<release>(?<release_v>\d+?\.\d+?\.\d+?)<\/release>\s+<api>(?<api_v>\d+?\.\d+?)<\/api>\s+<\/version>\s+<stability>\s+<release>(?<release_s>[a-z]+?)<\/release>\s+<api>(?<api_s>[a-z]+?)<\/api>\s+<\/stability>/';
+$package_ver_regex = '/<version>\s+<release>(?<release_v>\d+?\.\d+?\.\d+?(?:-?(?:alpine|beta|rc\d*?))?)<\/release>\s+<api>(?<api_v>\d+?\.\d+?)<\/api>\s+<\/version>\s+<stability>\s+<release>(?<release_s>[a-z]+?)<\/release>\s+<api>(?<api_s>[a-z]+?)<\/api>\s+<\/stability>/i';
 preg_match($package_ver_regex, file_get_contents(__DIR__ . '/../package.xml'), $matches);
 $package_release_ver = $matches['release_v'];
 $package_api_ver = $matches['api_v'];
 $package_release_stable = $matches['release_s'];
 $package_api_stable = $matches['api_s'];
 if (round($package_release_ver, 0, PHP_ROUND_HALF_DOWN) != $package_api_ver) {
-    error_exit("Wrong api version [{$package_api_ver}] with release version [{$package_release_ver}]");
+    swoole_error("Wrong api version [{$package_api_ver}] with release version [{$package_release_ver}]");
 }
 if ($package_release_stable . $package_api_stable !== 'stable' . 'stable') {
-    swoole_log(
-        "Notice: It's not a stable version, can't be released by pecl\n",
-        SWOOLE_COLOR_YELLOW
-    );
+    swoole_warn("It's not a stable version, can't be released by pecl.");
 }
 echo "[Version] => {$package_release_ver}\n";
 echo "[API-Ver] => {$package_api_ver}\n";
@@ -108,14 +123,13 @@ check_source_ver($package_release_ver, dirname(__DIR__) . '/include/swoole.h');
 check_source_ver($package_release_ver, dirname(__DIR__) . '/CMakeLists.txt');
 
 // check file lists
-$file_list_raw = `cd {$root_dir} && git ls-files`;
-$file_list_raw = explode("\n", $file_list_raw);
+$file_list_raw = swoole_git_files();
 $file_list = [];
 foreach ($file_list_raw as $file) {
     if (empty($file)) {
         continue;
     }
-    if (is_dir($root_dir . $file)) {
+    if (is_dir("{$root_dir}/{$file}")) {
         continue;
     }
     if ($file === 'package.xml' || substr($file, 0, 1) === '.') {
@@ -136,7 +150,13 @@ foreach ($file_list_raw as $file) {
                 $role = 'doc';
                 break;
             case '':
-                if (substr(file_get_contents($root_dir . $file), 0, 2) !== '#!') {
+                static $spacial_source_list = [
+                    'Makefile' => true
+                ];
+                if ($spacial_source_list[pathinfo($file, PATHINFO_BASENAME)] ?? false) {
+                    break;
+                }
+                if (substr(file_get_contents("{$root_dir}/{$file}"), 0, 2) !== '#!') {
                     $role = 'doc';
                 }
                 break;
@@ -147,7 +167,7 @@ foreach ($file_list_raw as $file) {
 
 $content = file_get_contents(__DIR__ . '/../package.xml');
 if (!preg_match('/([ ]*)\<dir[ ]name=\"\/\">/', $content, $matches)) {
-    error_exit('Match dir tag failed!');
+    swoole_error('Match dir tag failed!');
 }
 $space = strlen($matches[1]);
 $space += 4;
@@ -155,14 +175,25 @@ $space = str_repeat(' ', $space);
 $dir_tag = '<dir name="/">' . "\n";
 $content = preg_replace('/(\<dir[ ]name=\"\/\">)([\s\S]*?)(\n[ ]*?\<\/dir>)/', '$1$3', $content, 1, $success);
 if (!$success) {
-    error_exit('Replace old content failed!');
+    swoole_error('Replace old content failed!');
 }
 $content = str_replace($dir_tag, $dir_tag . $space . implode("{$space}", $file_list), $content, $success);
 if (!$success) {
-    error_exit('Replace new content failed!');
+    swoole_error('Replace new content failed!');
+}
+date_default_timezone_set('Asia/Shanghai');
+$date_tag = date('Y-m-d');
+$content = preg_replace('/(<date\>)\d+?-\d+?-\d+?(<\/date>)/', '${1}' . $date_tag . '${2}', $content, $success);
+if (!$success) {
+    swoole_error('Replace date tag failed!');
+}
+$time_tag = date('H', time() + 3600) . ':00:00';
+$content = preg_replace('/(<time\>)\d+?:\d+?:\d+?(<\/time>)/', '${1}' . $time_tag . '${2}', $content, $success);
+if (!$success) {
+    swoole_error('Replace time tag failed!');
 }
 if (!file_put_contents(__DIR__ . '/../package.xml', $content)) {
-    error_exit('Output package successful!');
+    swoole_error('Output package.xml failed!');
 }
 $package = trim(`cd {$root_dir} && pecl package`);
 if (preg_match('/Warning/i', $package)) {
@@ -171,8 +202,11 @@ if (preg_match('/Warning/i', $package)) {
     $warn = implode("\n", $warn);
     swoole_log("{$warn}\n", SWOOLE_COLOR_MAGENTA);
 }
-echo "{$package}\n";
 // check package status
-if (!preg_match('/Package swoole-[\d.]+\.tgz done/', $package)) {
-    error_exit();
+if (!preg_match('/Package (?<filename>swoole-.+?.tgz) done/', $package, $matches)) {
+    swoole_error($package);
+} else {
+    $file_name = $matches['filename'];
+    $file_size = file_size("{$root_dir}/{$file_name}");
+    swoole_success("Package {$file_name} ({$file_size}) done");
 }
